Cleansed the data, added KStar, Refactored Code
fixed the creation of instance, Rating Model updated to take decimal
values, changed queries to get distinct data, changed the dataset for
training and cleansed it.

+++ src/main/java/edu/ccsu/dao/RatingsJDBC.java
@@ -26,7 +26,7 @@ public class RatingsJDBC {
 
 		try {
 			jdbcTemplateObject.update(sql, new Object[] { rating.getUserName(), rating.getMovieID(),
-					Integer.parseInt(rating.getUserRating()) });
+					Double.parseDouble(rating.getUserRating()) });
 			return true;
 		} catch (Exception e) {
 			e.printStackTrace();
